requirement.txt
langchain_google_genai 
langgraph 
python-dotenv 
joblib 
pandas
"fastapi[standard]"
uvicorn